backend/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
scikit-learn
scipy
numpy
pandas 